Moved out of components folder
